# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Combo {
  id: number;
  name: string;
  price: number;
  hennaCount: number;
  nailCount: number;
  /** @nullable */
  qrCodeUrl?: string | null;
}

export interface CreateOrderBody {
  comboId: number;
  fullName: string;
  phone: string;
  whatsapp: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending_verification: "pending_verification",
  payment_verified: "payment_verified",
  order_confirmed: "order_confirmed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  comboId: number;
  comboName: string;
  price: number;
  fullName: string;
  phone: string;
  whatsapp: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  status: OrderStatus;
  createdAt: string;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending_verification: "pending_verification",
  payment_verified: "payment_verified",
  order_confirmed: "order_confirmed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface Analytics {
  totalOrders: number;
  todayOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  cancelledOrders: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetAdminOrdersParams = {
  search?: string;
  status?: string;
};