# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Al Mehandi e-commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: combos
    description: Combo offers
  - name: orders
    description: Customer orders
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /combos:
    get:
      operationId: getCombos
      tags: [combos]
      summary: List all combo offers
      responses:
        "200":
          description: List of combos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Combo"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID (for order tracking)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/orders:
    get:
      operationId: getAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/analytics:
    get:
      operationId: getAnalytics
      tags: [admin]
      summary: Get analytics data (admin)
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Combo:
      type: object
      required: [id, name, price, hennaCount, nailCount]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        hennaCount:
          type: integer
        nailCount:
          type: integer
        qrCodeUrl:
          type: ["string", "null"]

    CreateOrderBody:
      type: object
      required: [comboId, fullName, phone, whatsapp, address, city, state, pincode]
      properties:
        comboId:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        notes:
          type: ["string", "null"]

    Order:
      type: object
      required: [id, comboId, comboName, price, fullName, phone, whatsapp, address, city, state, pincode, status, createdAt]
      properties:
        id:
          type: integer
        comboId:
          type: integer
        comboName:
          type: string
        price:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        notes:
          type: ["string", "null"]
        paymentScreenshotUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending_verification, payment_verified, order_confirmed, cancelled]
        createdAt:
          type: string

    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AdminLoginResponse:
      type: object
      required: [success, token]
      properties:
        success:
          type: boolean
        token:
          type: string

    UpdateOrderStatusBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending_verification, payment_verified, order_confirmed, cancelled]

    Analytics:
      type: object
      required: [totalOrders, todayOrders, pendingOrders, confirmedOrders, cancelledOrders]
      properties:
        totalOrders:
          type: integer
        todayOrders:
          type: integer
        pendingOrders:
          type: integer
        confirmedOrders:
          type: integer
        cancelledOrders:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string