# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all combo offers
 */
export const GetCombosResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  hennaCount: zod.number(),
  nailCount: zod.number(),
  qrCodeUrl: zod.string().nullish(),
});
export const GetCombosResponse = zod.array(GetCombosResponseItem);

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  comboId: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get order by ID (for order tracking)
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  status: zod.enum([
    "pending_verification",
    "payment_verified",
    "order_confirmed",
    "cancelled",
  ]),
  createdAt: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
});

/**
 * @summary List all orders (admin)
 */
export const GetAdminOrdersQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const GetAdminOrdersResponseItem = zod.object({
  id: zod.number(),
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  status: zod.enum([
    "pending_verification",
    "payment_verified",
    "order_confirmed",
    "cancelled",
  ]),
  createdAt: zod.string(),
});
export const GetAdminOrdersResponse = zod.array(GetAdminOrdersResponseItem);

/**
 * @summary Update order status (admin)
 */
export const UpdateOrderStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateOrderStatusBody = zod.object({
  status: zod.enum([
    "pending_verification",
    "payment_verified",
    "order_confirmed",
    "cancelled",
  ]),
});

export const UpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  status: zod.enum([
    "pending_verification",
    "payment_verified",
    "order_confirmed",
    "cancelled",
  ]),
  createdAt: zod.string(),
});

/**
 * @summary Get analytics data (admin)
 */
export const GetAnalyticsResponse = zod.object({
  totalOrders: zod.number(),
  todayOrders: zod.number(),
  pendingOrders: zod.number(),
  confirmedOrders: zod.number(),
  cancelledOrders: zod.number(),
});