# lib/api-zod/src/generated/types/adminLoginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}