# lib/api-zod/src/generated/types/combo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Combo {
  id: number;
  name: string;
  price: number;
  hennaCount: number;
  nailCount: number;
  /** @nullable */
  qrCodeUrl?: string | null;
}