# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  comboId: number;
  comboName: string;
  price: number;
  fullName: string;
  phone: string;
  whatsapp: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  status: OrderStatus;
  createdAt: string;
}